Extract editstarts to handler.
